proyecto1.7/requirements.txt
polars
plotly
mysql-connector-python
python-dotenv
requests
pandas
numpy
pyarrow
